Project all CIGAR ops onto Pileup.Element.Alignment
This is a first fix for #30. It also re-enables the “read 5” test of
PileupSuite (#2).

The “long names” of `CigarOperator` are not enough
(c.f.
[net/sf/samtools/CigarOperator.java](https://github.com/nh13/picard/blob/master/src/java/net/sf/samtools/CigarOperator.java)
and
[documentation](http://picard.sourceforge.net/javadoc/net/sf/samtools/CigarOperator.html))
so I went back to the “one-letter” ones.

The assumptions are taken from this [wikipage about
SAM](http://davetang.org/wiki/tiki-index.php?page=SAM) and
Matt's
[blog post](http://zenfractal.com/2013/06/19/playing-with-matches/).